+++ packages/spruce-types/index.ts
@@ -5,6 +5,7 @@ export * from './src/events'
 export * from './src/events/reschedule-calendar-event'
 export * from './src/events/get-calendars'
 export * from './src/events/get-ui-enhancements'
+export * from './src/events/big-search'
 export * from './src/core'
 export * from './src/gql/queries'
 export { SpruceDataTypes } from './src/spruceDataTypes'
